closes #7, closes #8: removed output truncation and changed initialization in rspec
All ~100 rspec test were successful for CSL8 (OSV V8R0) and CSL9_V7R1

For a better initialization, up to 4 children are created, before the object is
destroyed recursively.

In addition, I have removed a truncation of the error messate to 1000 chars.
Instead, only part of the error message is shown per default on the Provisionings list.