.travis.yml: for testing, make it more similar to the one in travisCI branch
travisCI was successful in the travisCI branch, while it fails in the development branch
There are almost no differences in the development branch, so I have to figure out, what is wrong.
For that, I try to make the branches more alike. First action is to use the same .travis.yml file
content. For that, I will make changes in both branches.