closes #21: SystemSettings.method_missing: encapsulated self.where into a begin-rescue block
apart from a begin-resuce block aroung self.where database search, I have added an option in the if-block
that will return the mapped environment variable as boolean, if the table has not been created yet.

This change was needed to prevent a problem with rake db:migrate for the test database. See issue #21