added migration to add default admin user, if admin_user table is empty
It might happen that the db is empty (e.g. after a reset of the db).
In that case, the admin_user table is empty and it is not possible to log into Active Admin
This migration will make sure that a default admin@example.com with password=password is created in this case
The password should be changed immediately after that.