.travis.yml: added all tests again.
It seems like I have overlooked that the commit
https://github.com/oveits/ProvisioningEngine/commit/bbb71a19fd7cb2509af44931126d83f10e6e21b8
already has closed #27. Therefore, I need to reverse the troubleshooting steps.

However, I will keep the FactoryGirl test, that verifies that FactoryGirl.create(:admin_user) can
be called twice.

I have moved the "FactoryGirl should create admin user, if it does not exist and not raise an error, if the user exists already"
on top and out of the SystemSettings test block.