rspec init: removed all SystemSettings from test database
There was a problem with SystemSettings in the rspec tests: if the SystemSetting was
set in a previous rspec test, I could not control the SystemSetting by ENV variables anymore,
e.g. by editing config/application.yml. This will not be discovered by travisCI, since it does not
work with this config file and since the test DB is starting from scratch every time.
anymore by changing config/application.yml