closes #33 automated tests for issue #16: problem with synchronization in simulation mode
- added a synchronize test in async mode.

Notes:

First I have tried to run 'RAILS_ENV=test rake jobs:work', but the jobs created in the rspec did not show up.

I have found two possibilities to run Delayed::Jobs in the rspec:
1)  Delayed::Worker.delay_jobs = true
2)  Delayed::Worker.new.work_off

Both are running the Delayed Job in the foreground, and the issue #16 is not covered, since it is running
in the same memory space.

In order to simulate the job to be running in its own memory space, I have undefined manually all
HttpPostRequest class variables (especially the @@provisioned variable). For that, I have defined a new class
function named HttpPostRequest.remove_class_variables

Summary: in the rspec, I have added Delayed::Worker.new.work_off to run the jobs, and I have simulated the job
         to be run in its own memory space by clearing the class variables from memory with the new function
         HttpPostRequest.remove_class_variables