closes #16: read/write HttpPostRequest class variable @@provisioned from/to database
Even if in simulation mode, I expect an object to be in the provisioned state after
it is provisioned. In simulation mode, the provisioned state is captured in the
class variable @@provisioned.

The problem is, that in case of background jobs (Delayed Jobs), the HttpPostRequest
is running in its own memory space and always starts with an undefined @@provisioned
variable. Therefore, the real status is unknown to HttpPostRequest.

Here, I have used a new model PersistentHash to store and retrieve the value of @@provisioned.
With that measure, we have made sure that the real status is shared between foreground job
and all background jobs.

Additional changes:
- commented out unused methods in http_post_request.rb and set debug to false
  in lib/http_post_request.rb and in app/jobs/general_job.rb