In simulation mode, always read provisioned status from database
In addition, we clean the database ProvisionedHash entries for HttpPost @@provisioned class
variable before each test run in rspec. With that, we avoid the "exist already" error we had
reported in issue #37.

Later I have found out that this is not the right solution (to be corrected in the next commit.