closes #38: SystemSetting: changed default value to String "false" instead of boolean false
The values of SystemSetting variable are set to the same value as the corresponding environment variable
i.e. "true" or "false" in case of a boolean-like variable. If no corresponding environment variable
exists, we set the default to "false". By an error, I had set it to a boolean 'false', which is saved as "f"
instead of "false" in the table. This is corrected now.