Added SystemSetting.webportal_provisioningobjects_hide_inactivebuttons to ActiveAdmin portal
I have replaced
  ENV["WEBPORTAL_PROVISIONINGOBJECTS_HIDE_INACTIVEBUTTONS"]
by
  SystemSetting.webportal_provisioningobjects_hide_inactivebuttons
everywhere in the code. With that, the SystemSetting "WEBPORTAL_PROVISIONINGOBJECTS_HIDE_INACTIVEBUTTONS"
appears in the ActiveAdmin portal.

Note that a rails s restart of "rails server" was required after changing the code, but once this is accomplished, the
variable can be changed on the fly. The restart of the "rails server" was needed,
because some view elements (e.g. buttons) are defined
as functions, and functions are not re-read until the next restart.