closes #61: improved page reload via java script
Before, the page content erroneously was sporadically toggled back to the
previous page content as an error. iAs a resolution, we have re-organized the
the concept to search for elements with class=="refresh",
memorized the id (which must be unique for solving the problem)
periodically receive the full page from the server and replace
the DOM element with the corresponding id. If the id is unique
for the service, it can never happen, that an element of one page
is replaced by the content of the element of another page erroneously.

The new concept will help to allow a page elements refresh on pages,
which contain a form. Today, this is not possible, since a form
is misbehaving when a HTML element containing the form is replaced.
We now can assign unique IDs for smaller elements in the page, and
indicate with class="refresh" that it is to be updated.
This will bedone in a nedt step.