closes #62: refresh index pages
Now index pages like /customers, /sites and /users are refreshed every 5 sec.
This is controlled by a class="refress" marker in the page. Note that the
id of the HTML element must be unique within the whole web service.
Note also, that auto-refresh is skipped, if the the HTML element
has a form included, since forms do not behave correctly, if they are
refreshed.